Fix(results): Downloading callback paginated items
Fixes: #239
