+++ dev/use-cases/ecosounds.html
@@ -1,0 +1,79 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Web Components - Dani</title>
+    <script type="module" src="/src/components/index.ts"></script>
+  </head>
+
+  <body>
+    <header>
+      <h1>EcoSounds</h1>
+      <p>This is a header. Please test scrolling and downloading results.</p>
+
+      <p>
+        Warning This page contains some diagnostic audio samples. I strongly suggest that you
+        <u>do not play audio on this page.</u>
+      </p>
+    </header>
+
+    <oe-verification-grid id="verification-grid">
+      <oe-verification verified="true" shortcut="Y"></oe-verification>
+      <oe-verification verified="false" shortcut="N"></oe-verification>
+
+      <oe-data-source slot="data-source" for="verification-grid"></oe-data-source>
+    </oe-verification-grid>
+
+    <footer>
+      <p>This is a footer. Please test scrolling past the verification grid.</p>
+    </footer>
+
+    <script type="module">
+      function randomSample(dataset) {
+        return dataset[Math.floor(Math.random() * dataset.length)];
+      }
+
+      const possibleAudioSamples = [
+        "/example_1s.wav",
+        "/example.flac",
+        "/example2.flac",
+        "/120hz.wav",
+        "/diagnostic.wav",
+        "/merged_diagnostic.wav",
+      ];
+
+      const possibleTags = ["bird", "insect", "mammal", "reptile"];
+
+      function gridPaginator() {
+        const subjects = Array.from({ length: 25 }).map(() => {
+          const src = randomSample(possibleAudioSamples);
+          const tag = randomSample(possibleTags);
+
+          return { src, tag };
+        });
+        return { subjects };
+      }
+
+      function setup() {
+        const verificationGrid = document.querySelector("oe-verification-grid");
+        verificationGrid.getPage = gridPaginator;
+      }
+
+      window.addEventListener("load", () => setup());
+    </script>
+
+    <style>
+      body {
+        margin: 0;
+      }
+
+      header,
+      footer {
+        background-color: #333;
+        color: white;
+        padding: 1em;
+      }
+    </style>
+  </body>
+</html>
