+++ dev/use-cases/index.html
@@ -14,6 +14,7 @@
     <a href="/dev/use-cases/dani.html">Dani</a>
     <a href="/dev/use-cases/use-case-6.html">BirdNet + Additional Labeling Task</a>
     <a href="/dev/use-cases/dani-alternative.html">Dani (alternative)</a>
+    <a href="/dev/use-cases/ecosounds.html">EcoSounds</a>
 
     <style>
       a {
